Revert "feat: alert dialog when sign out"
This reverts commit 04afa4edde44e32049587731fc949ecdd81bad33.
