+++ web/next.config.js
@@ -3,7 +3,7 @@ const nextConfig = {
   reactStrictMode: true,
   swcMinify: true,
   images: {
-    domains: ["static-cdn.jtvnw.net", "cdn.discordapp.com"],
+    domains: ["static-cdn.jtvnw.net"],
   },
 };
