+++ web/package.json
@@ -11,10 +11,8 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^2.9.8",
-    "@radix-ui/react-alert-dialog": "^1.0.0",
     "@radix-ui/react-checkbox": "^1.0.0",
     "@radix-ui/react-dialog": "^1.0.0",
-    "@radix-ui/react-popover": "^1.0.0",
     "@radix-ui/react-select": "^1.0.0",
     "@radix-ui/react-toggle-group": "^1.0.0",
     "@supabase/supabase-js": "^1.35.7",
