+++ web/src/components/Form/Select.tsx
@@ -1,34 +1,17 @@
 import { useEffect, useState } from "react";
-import { toast } from "react-toastify";
 
 import { GameAd } from "@@types/GameAd";
 import * as SelectPrimitive from "@radix-ui/react-select";
 import styles from "@styles/select.module.css";
 import { CaretDown, CaretUp, Check } from "phosphor-react";
 
-import { supabase } from "@lib/supabase";
-
 export function Select() {
   const [gameAds, setGameAds] = useState<GameAd[]>([]);
 
   useEffect(() => {
-    (async () => {
-      try {
-        const { data } = await supabase
-          .from<GameAd>("Game")
-          .select("*")
-          .order("name", {
-            ascending: true,
-          })
-          .throwOnError();
-        if (data) {
-          setGameAds(data);
-        }
-      } catch (err) {
-        console.error(err);
-        toast.error("Ocorreu um erro ao buscar os jogos!");
-      }
-    })();
+    fetch("/api/games")
+      .then((res) => res.json())
+      .then(({ data }) => setGameAds(data));
   }, []);
 
   return (
