+++ web/src/components/Header.tsx
@@ -1,7 +1,21 @@
-import { ProfilePopover } from "./ProfilePopover";
+import { useAuth } from "@hooks/useAuth";
+import { supabase } from "@lib/supabase";
 
 /* eslint-disable @next/next/no-img-element */
 export function Header() {
+  const { user } = useAuth();
+
+  async function handleSignOut() {
+    try {
+      const { error } = await supabase.auth.signOut();
+      if (error) {
+        throw error;
+      }
+    } catch (err) {
+      console.error(err);
+    }
+  }
+
   return (
     <header className="flex items-center flex-col">
       <img
@@ -17,7 +31,11 @@ export function Header() {
         </span>{" "}
         está aqui.
       </h1>
-      <ProfilePopover />
+      {user && (
+        <div className="absolute top-0 right-0">
+          <button onClick={handleSignOut}>Sair</button>
+        </div>
+      )}
     </header>
   );
 }
